Further cleaning.
