+++ src/Nethermind.Int256.Test/UInt256Tests.cs
@@ -5,14 +5,14 @@ using NUnit.Framework;
 
 namespace Nethermind.Int256.Test
 {
-    public partial class UInt256Tests<T> where T : IInteger<T>
+    public partial class UInt256TestsTemplate<T> where T : IInteger<T>
     {
         protected readonly Func<BigInteger, T> convert;
         protected readonly Func<int, T> convertFromInt;
         protected readonly Func<BigInteger, BigInteger> postprocess;
         protected readonly BigInteger maxValue;
 
-        protected UInt256Tests(Func<BigInteger, T> convert, Func<int, T> convertFromInt, Func<BigInteger, BigInteger> postprocess, BigInteger maxValue)
+        protected UInt256TestsTemplate(Func<BigInteger, T> convert, Func<int, T> convertFromInt, Func<BigInteger, BigInteger> postprocess, BigInteger maxValue)
         {
             this.convert = convert;
             this.convertFromInt = convertFromInt;
@@ -214,9 +214,9 @@ namespace Nethermind.Int256.Test
         }
     }
 
-    public class UInt256ImplTests : UInt256Tests<UInt256>
+    public class UInt256Tests : UInt256TestsTemplate<UInt256>
     {
-        public UInt256ImplTests() : base((BigInteger x) => (UInt256)x, (int x) => (UInt256)x, x => x, TestNumbers.UInt256Max) { }
+        public UInt256Tests() : base((BigInteger x) => (UInt256)x, (int x) => (UInt256)x, x => x, TestNumbers.UInt256Max) { }
 
         [Test]
         public virtual void Zero_is_min_value()
@@ -255,75 +255,5 @@ namespace Nethermind.Int256.Test
         {
             convert(1).MaximalValue.Should().Be(convert(maxValue));
         }
-    }
-
-    public class Int256ImplTests : UInt256Tests<Int256>
-    {
-
-        private static BigInteger Postprocess(BigInteger big)
-        {
-            var bytes = big.ToByteArray();
-            return new BigInteger(bytes.AsSpan().Slice(0, Math.Min(256 / 8, bytes.Length)));
-        }
-
-        public Int256ImplTests() : base((BigInteger x) => new Int256(x), (int x) => new Int256(x), Postprocess, TestNumbers.Int256Max)
-        {
-        }
-
-        [TestCaseSource(typeof(BinaryOps), nameof(BinaryOps.SignedTestCases))]
-        public override void Add((BigInteger A, BigInteger B) test) => base.Add(test);
-
-        [TestCaseSource(typeof(TernaryOps), nameof(TernaryOps.SignedTestCases))]
-        public override void AddMod((BigInteger A, BigInteger B, BigInteger M) test) => base.AddMod(test);
-
-        [TestCaseSource(typeof(BinaryOps), nameof(BinaryOps.SignedTestCases))]
-        public override void Subtract((BigInteger A, BigInteger B) test) => base.Subtract(test);
-
-        [TestCaseSource(typeof(TernaryOps), nameof(TernaryOps.SignedTestCases))]
-        public void SubtractMod((BigInteger A, BigInteger B, BigInteger M) test)
-        {
-            if (test.M.IsZero)
-            {
-                return;
-            }
-
-            BigInteger resBigInt = ((test.A - test.B) % test.M) % (BigInteger.One << 256);
-            resBigInt = postprocess(resBigInt);
-
-            Int256 uint256a = convert(test.A);
-            Int256 uint256b = convert(test.B);
-            Int256 uint256m = convert(test.M);
-            uint256a.SubtractMod(uint256b, uint256m, out Int256 res);
-            res.Convert(out BigInteger resUInt256);
-
-            resUInt256.Should().Be(resBigInt);
-        }
-
-        [TestCaseSource(typeof(BinaryOps), nameof(BinaryOps.SignedTestCases))]
-        public override void Multiply((BigInteger A, BigInteger B) test) => base.Multiply(test);
-
-        [TestCaseSource(typeof(TernaryOps), nameof(TernaryOps.SignedTestCases))]
-        public override void MultiplyMod((BigInteger A, BigInteger B, BigInteger M) test) => base.MultiplyMod(test);
-
-        [TestCaseSource(typeof(BinaryOps), nameof(BinaryOps.SignedTestCases))]
-        public override void Div((BigInteger A, BigInteger B) test) => base.Div(test);
-
-        [TestCaseSource(typeof(BinaryOps), nameof(BinaryOps.SignedShiftTestCases))]
-        public override void Exp((BigInteger A, int n) test) => base.Exp(test);
-
-        [TestCaseSource(typeof(TernaryOps), nameof(TernaryOps.SignedTestCases))]
-        public override void ExpMod((BigInteger A, BigInteger B, BigInteger M) test) => base.ExpMod(test);
-
-        [TestCaseSource(typeof(BinaryOps), nameof(BinaryOps.SignedShiftTestCases))]
-        public override void Lsh((BigInteger A, int n) test) => base.Lsh(test);
-
-        [TestCaseSource(typeof(BinaryOps), nameof(BinaryOps.SignedShiftTestCases))]
-        public override void Rsh((BigInteger A, int n) test) => base.Rsh(test);
-
-        [TestCaseSource(typeof(UnaryOps), nameof(UnaryOps.SignedTestCases))]
-        public override void ToBigIntegerAndBack(BigInteger test) => base.ToBigIntegerAndBack(test);
-
-        [TestCaseSource(typeof(UnaryOps), nameof(UnaryOps.SignedTestCases))]
-        public override void ToString(BigInteger test) => base.ToString(test);
     }
 }
